# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Microblogging API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/{postId}/comments:
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
  /comments/{id}:
    patch:
      operationId: updateComment
      tags: [comments]
      summary: Update a comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentBody"
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List every category with its post count
      description: |
        Public read. Returns all categories ordered by name, each enriched with
        a `postCount` of *published* posts attached to it.
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriesList"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid body
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Slug already exists
  /categories/{slug}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithPostCount"
        "404":
          description: Not found
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category by id (owner only)
      description: |
        Writes are keyed by the stable internal `id` rather than the
        mutable `slug` so that renames (which change the slug) do not
        race against in-flight management UI requests. Express routes
        both `GET /categories/{slug}` and `PATCH/DELETE
        /categories/{id}` on the same path pattern; method dispatch is
        unambiguous.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid body
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "409":
          description: Slug already exists
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category by id (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /categories/{slug}/posts:
    get:
      operationId: getCategoryPosts
      tags: [categories]
      summary: List published posts attached to a category
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Category posts page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
        "404":
          description: Not found
...
components:
  schemas:
...
    Category:
      type: object
      description: Owner-managed taxonomy entry. Slug is the canonical identifier in URLs.
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - createdAt
        - updatedAt
    CategoryWithPostCount:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        postCount:
          type: integer
          description: Number of published posts attached to this category.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - postCount
        - createdAt
        - updatedAt
    CategoriesList:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryWithPostCount"
      required:
        - categories
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 191
          description: Optional. When omitted the server derives one from `name`.
        description:
          type: string
          nullable: true
      required:
        - name
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 191
        description:
          type: string
          nullable: true
    Post:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: string
          nullable: true
        content:
          type: string
        contentFormat:
          type: string
          enum: [plain, html]
        commentCount:
          type: integer
        sourceFeedId:
          type: integer
          nullable: true
          description: ID of the feed_sources row that imported this post (PESOS); null for owner-authored posts.
        sourceFeedName:
          type: string
          nullable: true
          description: Display name of the originating feed source (joined from feed_sources). Null for owner-authored posts.
        sourceCanonicalUrl:
          type: string
          nullable: true
          description: Permalink of the post on its origin site (PESOS attribution).
        categories:
          type: array
          description: Categories this post belongs to (owner-managed taxonomy).
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - authorId
        - authorName
        - content
        - contentFormat
        - commentCount
        - categories
        - createdAt
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: string
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorId
        - authorName
        - content
        - createdAt
...
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
          maxLength: 500
      required:
        - content
    UpdateCommentBody:
      type: object
      properties:
        content:
          type: string
          maxLength: 500
      required:
        - content
...